Update Lab.cpp
I worked on it a bit. The process now successfully creates files and Child Processes each time the user enters a string. WARNING: When the user enters "Done" an infinite loop is triggered idk why or how but I think the return EXIT_SUCCESS only terminates the Parent Process and not the children causing the child processes to go on infinitly.